// types/shared/alignment.ts
export enum HORIZONTAL_ALIGNMENT {
  LEFT = 'left',
  RIGHT = 'right',
  CENTER = 'center',
}

export enum VERTICAL_ALIGNMENT {
  TOP = 'top',
  BOTTOM = 'bottom',
  CENTER = 'center',
}

export enum TEXT_ALIGNMENT {
  LEFT = 'left',
  CENTER = 'center',
  RIGHT = 'right',
  JUSTIFY = 'justify',
}

export enum OBJECT_FIT {
  CONTAIN = 'contain',
  COVER = 'cover',
  FILL = 'fill',
}